refactor(channels): simplify channel rail composition
