+++ apps/web/src/features/channels-view/components/ChannelsMobileView.tsx
@@ -24,7 +24,7 @@ import { filterChannelsForScope } from '../queries';
 import type { ChannelsQueryScope } from '../types';
 import { channelMentionsUser } from '../utils';
 import { ChannelsEmptyState } from './ChannelsEmptyState';
-import { ConversationCard } from './rail/ConversationCard';
+import { ConversationCard } from './rail/ChannelRailItems';
 import { useChannelCalls } from './rail/hooks/useChannelCalls';
 import { useChannelRailActivity } from './rail/hooks/useChannelRailActivity';
